pyproject.toml
[tool.poetry]
name = "cps-childcare"
version = "0.0.1"
description = "Project to collect data on before and after school childcare programs at Chicago Public Schools."
authors = ["Michealngelo D'Agostino <mdagost@gmail.com>"]
readme = "README.md"
packages = [{include = "cps_childcare"}]
repository = "https://github.com/mdagost/cps-childcare"

# Requirements
[tool.poetry.dependencies]
python = "^3.10"
airtable-python-wrapper = "^0.15.3"
alembic = "^1.13.2"
firecrawl-py = "^1.2.3"
openai = "^1.43.0"
pandas = "^2.2.2"
pydantic = "^2.8.2"
sqlmodel = "^0.0.22"
tenacity = "^8.5.0"
tiktoken = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"